Corrige o bug com a VideoView
