+++ NSBrazilConf/Assets.xcassets/Colors/NSBRColor.colorset/Contents.json
@@ -9,10 +9,10 @@
       "color" : {
         "color-space" : "srgb",
         "components" : {
-          "red" : "0xC0",
+          "red" : "0x1E",
           "alpha" : "1.000",
-          "blue" : "0xFE",
-          "green" : "0xEF"
+          "blue" : "0x29",
+          "green" : "0x20"
         }
       }
     }
