+++ NSBrazilConf/Assets.xcassets/Colors/Orange.colorset/Contents.json
