+++ NSBrazilConf/Assets.xcassets/Colors/abColor.colorset/Contents.json
@@ -1,0 +1,20 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0.753",
+          "alpha" : "1.000",
+          "blue" : "0.996",
+          "green" : "0.937"
+        }
+      }
+    }
+  ]
+}
