+++ NSBrazilConf/Assets.xcassets/Colors/chColor.colorset/Contents.json
