+++ NSBrazilConf/Assets.xcassets/Colors/iFoodColor.colorset/Contents.json
@@ -9,10 +9,10 @@
       "color" : {
         "color-space" : "srgb",
         "components" : {
-          "red" : "0xFE",
+          "red" : "0.929",
           "alpha" : "1.000",
-          "blue" : "0x01",
-          "green" : "0xE8"
+          "blue" : "0.133",
+          "green" : "0.090"
         }
       }
     }
