+++ NSBrazilConf/Assets.xcassets/Colors/mcColor.colorset/Contents.json
@@ -9,10 +9,10 @@
       "color" : {
         "color-space" : "srgb",
         "components" : {
-          "red" : "0xED",
+          "red" : "0.996",
           "alpha" : "1.000",
-          "blue" : "0x22",
-          "green" : "0x17"
+          "blue" : "0.004",
+          "green" : "0.910"
         }
       }
     }
